[SECURITY-897] Fold tests back into normal files
